+++ src/main/java/com/travel/domain/follow/entity/Follow.java
@@ -1,6 +1,6 @@
 package com.travel.domain.follow.entity;
 
-import com.travel.domain.member.domain.Member;
+import com.travel.domain.member.entity.Member;
 import com.travel.global.common.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.Getter;
