+++ src/main/java/com/travel/domain/member/dao/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.travel.domain.member.dao;
 
-import com.travel.domain.member.domain.Member;
+import com.travel.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
