+++ src/main/java/com/travel/domain/member/dao/RefreshTokenRepository.java
@@ -1,6 +1,6 @@
 package com.travel.domain.member.dao;
 
-import com.travel.domain.member.domain.RefreshToken;
+import com.travel.domain.member.entity.RefreshToken;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
