+++ src/main/java/com/travel/domain/member/entity/MemberRole.java
@@ -1,4 +1,4 @@
-package com.travel.domain.member.domain;
+package com.travel.domain.member.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
