+++ src/main/java/com/travel/domain/member/entity/Provider.java
@@ -1,4 +1,4 @@
-package com.travel.domain.member.domain;
+package com.travel.domain.member.entity;
 
 import lombok.Getter;
