+++ src/main/java/com/travel/domain/tag/entity/Tag.java
@@ -1,7 +1,7 @@
-package com.travel.domain.tag.domain;
+package com.travel.domain.tag.entity;
 
-import com.travel.domain.place.domain.Place;
-import com.travel.domain.tagtype.domain.TagType;
+import com.travel.domain.place.entity.Place;
+import com.travel.domain.tagtype.entity.TagType;
 import com.travel.global.common.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.Getter;
