+++ src/main/java/com/travel/security/auth/dto/token/AccessTokenDto.java
@@ -1,5 +1,5 @@
 package com.travel.security.auth.dto.token;
 
-import com.travel.domain.member.domain.MemberRole;
+import com.travel.domain.member.entity.MemberRole;
 
 public record AccessTokenDto(Long memberId, MemberRole memberRole, String token) {}
